+++ examples/interconnects.tf
@@ -1,0 +1,15 @@
+provider "oneview" {
+	ov_username = "<username>"
+	ov_password = "<password>"
+	ov_endpoint = "<endpoint>"
+	ov_sslverify = false
+	ov_apiversion = 800
+}
+
+data "oneview_interconnect" "interconnect" {
+	name = "EN1, interconnect 3"
+}
+
+output "oneiew_interconnect_value" {
+	value = "${data.oneview_interconnect.interconnect.uri}"
+}