+++ oneview/provider.go
@@ -98,14 +98,21 @@ func Provider() terraform.ResourceProvider {
 		},
 
 		DataSourcesMap: map[string]*schema.Resource{
-			"oneview_scope":           dataSourceScope(),
-			"oneview_server_hardware": dataSourceServerHardware(),
-			"oneview_server_profile":  dataSourceServerProfile(),
-		},
+			"oneview_interconnect_type":    dataSourceInterconnectType(),
+			"oneview_interconnect":         dataSourceInterconnects(),
+			"oneview_logical_interconnect": dataSourceLogicalInterconnect(),
+			"oneview_scope":                dataSourceScope(),
+			"oneview_server_hardware":      dataSourceServerHardware(),
+			"oneview_server_hardware_type": dataSourceServerHardwareType(),
+			"oneview_logical_enclosure":    dataSourceLogicalEnclosure(),
+			"oneview_enclosure_group":      dataSourceEnclosureGroup(),
+ 			"oneview_server_profile":  dataSourceServerProfile(),
+    },
 
 		ResourcesMap: map[string]*schema.Resource{
 			"oneview_server_profile":             resourceServerProfile(),
 			"oneview_enclosure":                  resourceEnclosure(),
+			"oneview_enclosure_group":            resourceEnclosureGroup(),
 			"oneview_ethernet_network":           resourceEthernetNetwork(),
 			"oneview_network_set":                resourceNetworkSet(),
 			"oneview_fcoe_network":               resourceFCoENetwork(),
@@ -117,6 +124,7 @@ func Provider() terraform.ResourceProvider {
 			"oneview_uplink_set":                 resourceUplinkSet(),
 			"oneview_icsp_server":                resourceIcspServer(),
 			"oneview_i3s_plan":                   resourceI3SPlan(),
+			"oneview_logical_enclosure":          resourceLogicalEnclosure(),
 		},
 		ConfigureFunc: providerConfigure,
 	}