+++ CHANGELOG.md
@@ -20,10 +20,16 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - [#47] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/47) Added Synergy Support for Logical Interconnect Groups
 
 ### Oneview Features supported
-- Enclosure    
+- Enclosure
+- Enclosure group
 - FC network
+- Interconnect
+- Interconnect type
+- Logical enclosure
+- Logical interconnect
 - Logical interconnect group
 - Scope
 - Server hardware
+- Server hardware type
 - Server profile
-- Uplink Set
+- Uplink set